Refactor CreateBuildsRequestProcessors
They shared a lot of common code so lets use a common method. Makes changing it a lot easier. Also makes FactoryManager.EventFromServer and FactoryManager.EventFromClient more consistent with usage in the rest of the mod: Factory.EventFromServer occurs when a client is processing an event from the server and Factory.EventFromClient when the server is processing an event from the client.

Now we should be able to rely on these two conditions along with packet author ids for most checks.

Might fix race condition related to client and host building belt at the same exact time.