PlayerAction_Build_Patch: Potential fix for errors when landing on planet that has a player actively building
The issue seems to arise from our setting of the factory references in Nebula happening at the same time that the game changes references upon arrival at a planet. Depending on the timing that this occurs, perhaps simply allowing the game to set references when the target planet is the same as the player's planet will solve this.